Added forum functionality for Louisiana Tech University.
Navigate to forum by pressing the new button on the index. Small bug where creating a post doesn't show up and you have to go to main screen and then back to forum for it to appear. Creating a new thread works fine though. TODO: replace the "createdBy" section when making a new post/thread to be the first and last name user used to sign up.