Added follow/unfollow and following screen
Added the feature for users to be able to follow/unfollow subgroups, added a new screen titled "FollowedForums" which displays the subgroups that the user is following. Tapping on "view" navigates the user to that specific subgroup through the "ColForum" screen. The college that the subgroup is contained in is also displayed on the following screen to hopefully help in avoiding confusion.