fix: email link sign, however:
Important Note: Dynamic Links are no longer supported in Firebase and will stop working August 25, 2025.

Email links now reliably get sent to the email address, and user sign in is consistently effective. However, it seems due to the way emulators handle links ("lack of native integration"), clicking the link will bring you to a webpage error of "handleCodeInApp must be set to true." This is set to true in the code so it should redirect the user to the app as expected when using a real android device. Lastly and also due to the flawed redirect, to sign in/create an account on the emulator via email link the user must click the link they receive and then reload the app. Extra: Code related to Twitter log in has been removed due to lack of support from react-native-firebase and it still having issues even after attempting to implement it with a third-party libraries and authenticators (auth0).