feat: added logout to settings screen
User can logout by clicking the button in the settings screen. The app also checks to see if a user is signed in before rendering the Launch or Home screen. If user is already logged in upon loading the app, go straight to the Home screen.
Extra: The phone verification works with the new navigation system now too.